LF-4696 Move signal-01.svg
